refactor(resource): move the inline test modules to siblings
`resource.rs` was 3,678 lines, 926 of them inline tests. Both modules move
to sibling files, which is what `router.rs`, `transport/http.rs`,
`tool.rs`, and `client/mod.rs` already do, and leaves the file at 2,754.

They have to be sibling modules rather than integration tests under
`tests/`: they reach private items such as `compile_uri_template` and
`decode_query_component`, which an integration test cannot see.

Pure movement. The retained body of `resource.rs` is byte-identical, and
each extracted file is byte-identical to the block it came from apart
from the uniform four-space dedent. All 47 tests still run.

Part of #1256.
