+++ crates/tower-mcp/src/resource.rs
@@ -2748,931 +2748,7 @@ fn extract_query_variables(query: &str, declared: &[String]) -> HashMap<String, 
 }
 
 #[cfg(test)]
-mod query_expansion_tests {
-    use super::*;
-
-    fn template(pattern: &str) -> ResourceTemplate {
-        ResourceTemplateBuilder::new(pattern).name("t").handler(
-            |uri: String, _vars: HashMap<String, String>| async move {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri,
-                        mime_type: None,
-                        text: Some("body".into()),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            },
-        )
-    }
-
-    /// #1282: a template could not carry the metadata a plain resource can,
-    /// and `definition()` emitted `arguments: []` unconditionally.
-    ///
-    /// The clone is the point of the assertion. `ResourceTemplate` implements
-    /// `Clone` by hand because of the `Arc<dyn ..>` handler, so a new field
-    /// that is not added there is dropped by every router that stores one.
-    #[test]
-    fn arguments_and_meta_reach_the_definition_and_survive_a_clone() {
-        let template = ResourceTemplateBuilder::new("db://{table}{?limit}")
-            .name("rows")
-            .argument("table", Some("Table to read"), true)
-            .argument("limit", None::<String>, false)
-            .handler(|uri: String, _vars: HashMap<String, String>| async move {
-                Ok(ReadResourceResult::text(uri, "[]"))
-            })
-            .with_meta(serde_json::json!({ "example.com/tier": "internal" }))
-            .expect("a valid _meta key");
-
-        for (label, definition) in [
-            ("original", template.definition()),
-            ("clone", template.clone().definition()),
-        ] {
-            assert_eq!(definition.arguments.len(), 2, "{label} lost its arguments");
-            assert_eq!(definition.arguments[0].name, "table");
-            assert_eq!(
-                definition.arguments[0].description.as_deref(),
-                Some("Table to read")
-            );
-            assert!(definition.arguments[0].required);
-            assert!(!definition.arguments[1].required);
-            assert!(definition.arguments[1].description.is_none());
-            assert_eq!(
-                definition
-                    .meta
-                    .as_ref()
-                    .and_then(|m| m.get("example.com/tier")),
-                Some(&serde_json::json!("internal")),
-                "{label} lost its _meta"
-            );
-        }
-    }
-
-    /// A template that declares nothing still advertises nothing, so the wire
-    /// output is unchanged for every server that has not opted in.
-    #[test]
-    fn a_template_without_declared_arguments_advertises_none() {
-        let definition = template("db://users/{id}").definition();
-        assert!(definition.arguments.is_empty());
-        assert!(definition.meta.is_none());
-    }
-
-    /// #1253: `{?cursor,limit}` was read as one variable literally named
-    /// "?cursor,limit", with a required capture, so the base URI did not
-    /// match at all.
-    #[test]
-    fn the_base_uri_matches_when_no_query_variable_is_present() {
-        let t = template("agent://threads{?cursor,limit}");
-        let vars = t.match_uri("agent://threads").expect("base URI must match");
-        assert!(
-            vars.is_empty(),
-            "no query variables were supplied: {vars:?}"
-        );
-    }
-
-    #[test]
-    fn each_query_variable_arrives_under_its_own_name() {
-        let t = template("agent://threads{?cursor,limit}");
-        let vars = t
-            .match_uri("agent://threads?cursor=abc&limit=20")
-            .expect("must match");
-        assert_eq!(vars.get("cursor").map(String::as_str), Some("abc"));
-        assert_eq!(vars.get("limit").map(String::as_str), Some("20"));
-    }
-
-    /// Every declared variable is optional and order is not significant,
-    /// which is the routing policy RFC 6570 leaves open.
-    #[test]
-    fn any_subset_in_any_order_matches() {
-        let t = template("codex://threads{?cursor,limit,cwd}");
-        let vars = t
-            .match_uri("codex://threads?cwd=%2Ftmp&cursor=x")
-            .expect("must match");
-        assert_eq!(vars.get("cursor").map(String::as_str), Some("x"));
-        assert_eq!(vars.get("cwd").map(String::as_str), Some("/tmp"));
-        assert!(!vars.contains_key("limit"), "absent stays absent: {vars:?}");
-    }
-
-    #[test]
-    fn query_expansion_composes_with_path_variables() {
-        let t = template("claude://projects/{project_key}/sessions{?cursor}");
-        let vars = t
-            .match_uri("claude://projects/abc/sessions?cursor=n2")
-            .expect("must match");
-        assert_eq!(vars.get("project_key").map(String::as_str), Some("abc"));
-        assert_eq!(vars.get("cursor").map(String::as_str), Some("n2"));
-
-        let bare = t
-            .match_uri("claude://projects/abc/sessions")
-            .expect("must match without a query");
-        assert_eq!(bare.get("project_key").map(String::as_str), Some("abc"));
-        assert!(!bare.contains_key("cursor"));
-    }
-
-    /// Present-but-empty is a different fact from absent, and a handler that
-    /// distinguishes them should be able to.
-    #[test]
-    fn an_empty_value_is_present_not_absent() {
-        let t = template("agent://threads{?cursor}");
-        let vars = t.match_uri("agent://threads?cursor=").expect("must match");
-        assert_eq!(vars.get("cursor").map(String::as_str), Some(""));
-    }
-
-    /// Routing must not break because a caller appended something we never
-    /// declared, and a repeated key resolves predictably.
-    #[test]
-    fn undeclared_keys_are_ignored_and_the_first_duplicate_wins() {
-        let t = template("agent://threads{?cursor}");
-        let vars = t
-            .match_uri("agent://threads?utm=x&cursor=first&cursor=second")
-            .expect("must match");
-        assert_eq!(vars.get("cursor").map(String::as_str), Some("first"));
-        assert!(!vars.contains_key("utm"));
-    }
-
-    #[test]
-    fn values_are_percent_decoded_and_plus_is_a_space() {
-        let t = template("agent://search{?q}");
-        let vars = t
-            .match_uri("agent://search?q=a%20b+c%2Fd")
-            .expect("must match");
-        assert_eq!(vars.get("q").map(String::as_str), Some("a b c/d"));
-    }
-
-    /// A percent sign next to a multibyte character used to panic: the
-    /// decoder sliced the string by byte offset after checking only length,
-    /// so either end of the slice could land inside a character. Any client
-    /// could reach it through a query expression, because even an undeclared
-    /// key is decoded before it is discarded.
-    #[test]
-    fn percent_decoding_survives_multibyte_input() {
-        let t = template("agent://search{?q}");
-        for (uri, expected) in [
-            ("agent://search?q=%a\u{e9}", "%a\u{e9}"),
-            ("agent://search?q=a%9\u{e9}", "a%9\u{e9}"),
-            ("agent://search?q=\u{e9}%\u{e9}", "\u{e9}%\u{e9}"),
-            ("agent://search?q=%\u{1f600}", "%\u{1f600}"),
-        ] {
-            let vars = t
-                .match_uri(uri)
-                .unwrap_or_else(|| panic!("must match: {uri}"));
-            assert_eq!(
-                vars.get("q").map(String::as_str),
-                Some(expected),
-                "for {uri}"
-            );
-        }
-    }
-
-    /// An undeclared key is decoded before it is discarded, so it is just as
-    /// reachable as a declared one.
-    #[test]
-    fn an_undeclared_key_with_multibyte_input_is_also_safe() {
-        let t = template("agent://search{?q}");
-        let vars = t
-            .match_uri("agent://search?%a\u{e9}=x&q=fine")
-            .expect("must match");
-        assert_eq!(vars.get("q").map(String::as_str), Some("fine"));
-    }
-
-    /// A percent at or near the end has no two digits to read.
-    #[test]
-    fn a_truncated_escape_is_left_alone() {
-        let t = template("agent://search{?q}");
-        for (uri, expected) in [
-            ("agent://search?q=%", "%"),
-            ("agent://search?q=%2", "%2"),
-            ("agent://search?q=a%", "a%"),
-            ("agent://search?q=%zz", "%zz"),
-        ] {
-            let vars = t.match_uri(uri).expect("must match");
-            assert_eq!(
-                vars.get("q").map(String::as_str),
-                Some(expected),
-                "for {uri}"
-            );
-        }
-    }
-
-    /// A malformed escape reaches the handler as written rather than
-    /// disappearing, so the caller can see what they sent.
-    #[test]
-    fn a_malformed_escape_is_left_alone() {
-        let t = template("agent://search{?q}");
-        let vars = t.match_uri("agent://search?q=100%").expect("must match");
-        assert_eq!(vars.get("q").map(String::as_str), Some("100%"));
-    }
-
-    /// The `&` continuation form is the same expansion.
-    #[test]
-    fn the_continuation_operator_is_supported() {
-        let t = template("agent://threads{&cursor}");
-        let vars = t.match_uri("agent://threads?cursor=z").expect("must match");
-        assert_eq!(vars.get("cursor").map(String::as_str), Some("z"));
-    }
-
-    /// A template without a query expression keeps matching a literal `?`
-    /// exactly as before, which is what stops this being a breaking change.
-    #[test]
-    fn templates_without_a_query_expression_are_unchanged() {
-        let compiled = compile_uri_template("http://example.com/api?query={q}").unwrap();
-        assert!(compiled.query_variables.is_empty());
-        assert_eq!(compiled.variables, vec!["q".to_string()]);
-
-        let t = template("http://example.com/api?query={q}");
-        let vars = t
-            .match_uri("http://example.com/api?query=hello")
-            .expect("must match");
-        assert_eq!(vars.get("q").map(String::as_str), Some("hello"));
-    }
-
-    #[test]
-    fn a_query_expression_must_end_the_template() {
-        assert!(compile_uri_template("agent://x{?a}/more").is_err());
-        assert!(compile_uri_template("agent://x{?a}{?b}").is_err());
-        assert!(compile_uri_template("agent://x{?}").is_err());
-        assert!(compile_uri_template("agent://x{?a,}").is_err());
-    }
-}
+mod query_expansion_tests;
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::time::Duration;
-    use tower::timeout::TimeoutLayer;
-
-    #[tokio::test]
-    async fn test_builder_resource() {
-        let resource = ResourceBuilder::new("file:///test.txt")
-            .name("Test File")
-            .description("A test file")
-            .text("Hello, World!");
-
-        assert_eq!(resource.uri, "file:///test.txt");
-        assert_eq!(resource.name, "Test File");
-        assert_eq!(resource.description.as_deref(), Some("A test file"));
-
-        let result = resource.read().await;
-        assert_eq!(result.contents.len(), 1);
-        assert_eq!(result.contents[0].text.as_deref(), Some("Hello, World!"));
-    }
-
-    #[cfg(feature = "stateless")]
-    #[tokio::test]
-    async fn test_mrtr_builder_preserves_input_required_outcome() {
-        let resource = ResourceBuilder::new("test://continue")
-            .mrtr_handler(|_ctx| async move {
-                Ok(RequestOutcome::input_required(
-                    crate::protocol::InputRequiredResult::new().with_request_state("signed-state"),
-                ))
-            })
-            .build();
-
-        let outcome = resource
-            .read_outcome_with_context(RequestContext::new(crate::protocol::RequestId::Number(1)))
-            .await
-            .unwrap();
-        assert_eq!(
-            outcome
-                .as_input_required()
-                .and_then(|result| result.request_state.as_deref()),
-            Some("signed-state")
-        );
-    }
-
-    // =========================================================================
-    // Clone vs. handler kind (#1340)
-    // =========================================================================
-    //
-    // `Resource::clone` (see the hand-written `impl Clone for Resource` above)
-    // is exercised incidentally by `read()`/`read_with_context()`, which clone
-    // `self` before dispatching -- so a plain-handler resource's clone was
-    // already covered by every test that calls `.read()`. `read_outcome_with_context`,
-    // which the router actually calls, does not clone `self`, only the handler
-    // behind it, so nothing here had exercised `Resource::clone` for an MRTR
-    // handler at all. `Tool::clone` dropping a field (#1298) is the reason
-    // this gets a dedicated, named test rather than staying implicit.
-
-    #[tokio::test]
-    async fn resource_clone_carries_the_plain_handler() {
-        let resource = ResourceBuilder::new("memory://cloneable")
-            .name("Cloneable")
-            .handler(|| async {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri: "memory://cloneable".to_string(),
-                        mime_type: Some("text/plain".to_string()),
-                        text: Some("original".to_string()),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            })
-            .build();
-
-        let cloned = resource.clone();
-        let result = cloned.read().await;
-        assert_eq!(result.contents[0].text.as_deref(), Some("original"));
-    }
-
-    #[cfg(feature = "stateless")]
-    #[tokio::test]
-    async fn resource_clone_carries_the_mrtr_handler() {
-        let resource = ResourceBuilder::new("test://cloneable-continue")
-            .mrtr_handler(|_ctx| async move {
-                Ok(RequestOutcome::input_required(
-                    crate::protocol::InputRequiredResult::new().with_request_state("cloned-state"),
-                ))
-            })
-            .build();
-
-        let cloned = resource.clone();
-        let outcome = cloned
-            .read_outcome_with_context(RequestContext::new(crate::protocol::RequestId::Number(1)))
-            .await
-            .unwrap();
-        assert_eq!(
-            outcome
-                .as_input_required()
-                .and_then(|result| result.request_state.as_deref()),
-            Some("cloned-state"),
-            "a dropped mrtr_handler after clone would fall through to the \
-             absent `service` and panic on `.expect(...)` instead of \
-             returning this"
-        );
-    }
-
-    #[cfg(feature = "stateless")]
-    #[tokio::test]
-    async fn mrtr_resource_composes_middleware() {
-        let resource = ResourceBuilder::new("test://layered-continue")
-            .mrtr_handler(|_ctx| async move {
-                Ok(RequestOutcome::input_required(
-                    crate::protocol::InputRequiredResult::new().with_request_state("layered-state"),
-                ))
-            })
-            .layer(TimeoutLayer::new(Duration::from_secs(1)))
-            .build();
-
-        let outcome = resource
-            .read_outcome_with_context(RequestContext::new(crate::protocol::RequestId::Number(2)))
-            .await
-            .unwrap();
-        assert_eq!(
-            outcome
-                .as_input_required()
-                .and_then(|result| result.request_state.as_deref()),
-            Some("layered-state")
-        );
-    }
-
-    #[cfg(feature = "stateless")]
-    #[tokio::test]
-    async fn test_mrtr_template_preserves_context_and_input_required_outcome() {
-        let template = ResourceTemplateBuilder::new("test://items/{id}").mrtr_handler(
-            |ctx, uri, variables| async move {
-                assert_eq!(ctx.request_state(), Some("prior-state"));
-                assert_eq!(uri, "test://items/42");
-                assert_eq!(variables.get("id").map(String::as_str), Some("42"));
-                Ok(RequestOutcome::input_required(
-                    crate::protocol::InputRequiredResult::new().with_request_state("next-state"),
-                ))
-            },
-        );
-        let variables = template.match_uri("test://items/42").unwrap();
-        let mut ctx = RequestContext::new(crate::protocol::RequestId::Number(1));
-        ctx.extensions_mut().insert(crate::mrtr::MrtrRequest::new(
-            None,
-            Some("prior-state".into()),
-        ));
-
-        let outcome = template
-            .read_outcome_with_context(ctx, "test://items/42", variables)
-            .await
-            .unwrap();
-        assert_eq!(
-            outcome
-                .as_input_required()
-                .and_then(|result| result.request_state.as_deref()),
-            Some("next-state")
-        );
-    }
-
-    #[tokio::test]
-    async fn test_json_resource() {
-        let resource = ResourceBuilder::new("file:///config.json")
-            .name("Config")
-            .json(serde_json::json!({"key": "value"}));
-
-        assert_eq!(resource.mime_type.as_deref(), Some("application/json"));
-
-        let result = resource.read().await;
-        assert!(result.contents[0].text.as_ref().unwrap().contains("key"));
-    }
-
-    #[tokio::test]
-    async fn test_handler_resource() {
-        let resource = ResourceBuilder::new("memory://counter")
-            .name("Counter")
-            .handler(|| async {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri: "memory://counter".to_string(),
-                        mime_type: Some("text/plain".to_string()),
-                        text: Some("42".to_string()),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            })
-            .build();
-
-        let result = resource.read().await;
-        assert_eq!(result.contents[0].text.as_deref(), Some("42"));
-    }
-
-    #[tokio::test]
-    async fn test_handler_resource_with_layer() {
-        let resource = ResourceBuilder::new("file:///with-timeout.txt")
-            .name("Resource with Timeout")
-            .handler(|| async {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri: "file:///with-timeout.txt".to_string(),
-                        mime_type: Some("text/plain".to_string()),
-                        text: Some("content".to_string()),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            })
-            .layer(TimeoutLayer::new(Duration::from_secs(30)))
-            .build();
-
-        let result = resource.read().await;
-        assert_eq!(result.contents[0].text.as_deref(), Some("content"));
-    }
-
-    #[tokio::test]
-    async fn test_handler_resource_with_timeout_error() {
-        let resource = ResourceBuilder::new("file:///slow.txt")
-            .name("Slow Resource")
-            .handler(|| async {
-                // Sleep much longer than timeout to ensure timeout fires reliably in CI
-                tokio::time::sleep(Duration::from_secs(1)).await;
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri: "file:///slow.txt".to_string(),
-                        mime_type: Some("text/plain".to_string()),
-                        text: Some("content".to_string()),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            })
-            .layer(TimeoutLayer::new(Duration::from_millis(50)))
-            .build();
-
-        let result = resource.read().await;
-        // read() still renders an error as content, since its signature has
-        // no room for a Result. The timeout is a genuine middleware
-        // failure (#1280), so it is sanitized to -32603 rather than
-        // leaking the raw middleware error.
-        assert!(result.contents[0].text.as_ref().unwrap().contains("-32603"));
-    }
-
-    #[tokio::test]
-    async fn test_context_aware_handler() {
-        let resource = ResourceBuilder::new("file:///ctx.txt")
-            .name("Context Resource")
-            .handler_with_context(|_ctx: RequestContext| async {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri: "file:///ctx.txt".to_string(),
-                        mime_type: Some("text/plain".to_string()),
-                        text: Some("context aware".to_string()),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            })
-            .build();
-
-        let result = resource.read().await;
-        assert_eq!(result.contents[0].text.as_deref(), Some("context aware"));
-    }
-
-    #[tokio::test]
-    async fn test_context_aware_handler_with_layer() {
-        let resource = ResourceBuilder::new("file:///ctx-layer.txt")
-            .name("Context Resource with Layer")
-            .handler_with_context(|_ctx: RequestContext| async {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri: "file:///ctx-layer.txt".to_string(),
-                        mime_type: Some("text/plain".to_string()),
-                        text: Some("context with layer".to_string()),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            })
-            .layer(TimeoutLayer::new(Duration::from_secs(30)))
-            .build();
-
-        let result = resource.read().await;
-        assert_eq!(
-            result.contents[0].text.as_deref(),
-            Some("context with layer")
-        );
-    }
-
-    #[tokio::test]
-    async fn test_trait_resource() {
-        struct TestResource;
-
-        impl McpResource for TestResource {
-            const URI: &'static str = "test://resource";
-            const NAME: &'static str = "Test";
-            const DESCRIPTION: Option<&'static str> = Some("A test resource");
-            const MIME_TYPE: Option<&'static str> = Some("text/plain");
-
-            async fn read(&self) -> Result<ReadResourceResult> {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri: Self::URI.to_string(),
-                        mime_type: Self::MIME_TYPE.map(|s| s.to_string()),
-                        text: Some("test content".to_string()),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            }
-        }
-
-        let resource = TestResource.into_resource();
-        assert_eq!(resource.uri, "test://resource");
-        assert_eq!(resource.name, "Test");
-
-        let result = resource.read().await;
-        assert_eq!(result.contents[0].text.as_deref(), Some("test content"));
-    }
-
-    #[test]
-    fn test_resource_definition() {
-        let resource = ResourceBuilder::new("file:///test.txt")
-            .name("Test")
-            .description("Description")
-            .mime_type("text/plain")
-            .text("content");
-
-        let def = resource.definition();
-        assert_eq!(def.uri, "file:///test.txt");
-        assert_eq!(def.name, "Test");
-        assert_eq!(def.description.as_deref(), Some("Description"));
-        assert_eq!(def.mime_type.as_deref(), Some("text/plain"));
-    }
-
-    #[test]
-    fn test_resource_request_new() {
-        let ctx = RequestContext::new(crate::protocol::RequestId::Number(1));
-        let req = ResourceRequest::new(ctx, "file:///test.txt".to_string());
-        assert_eq!(req.uri, "file:///test.txt");
-    }
-
-    #[test]
-    fn test_resource_catch_error_clone() {
-        let handler = FnHandler {
-            handler: || async {
-                Ok::<_, Error>(ReadResourceResult {
-                    contents: vec![],
-                    meta: None,
-                    ..Default::default()
-                })
-            },
-        };
-        let service = ResourceHandlerService::new(handler);
-        let catch_error = ResourceCatchError::new(service);
-        let _clone = catch_error.clone();
-    }
-
-    #[test]
-    fn test_resource_catch_error_debug() {
-        let handler = FnHandler {
-            handler: || async {
-                Ok::<_, Error>(ReadResourceResult {
-                    contents: vec![],
-                    meta: None,
-                    ..Default::default()
-                })
-            },
-        };
-        let service = ResourceHandlerService::new(handler);
-        let catch_error = ResourceCatchError::new(service);
-        let debug = format!("{:?}", catch_error);
-        assert!(debug.contains("ResourceCatchError"));
-    }
-
-    // =========================================================================
-    // Resource Template Tests
-    // =========================================================================
-
-    #[test]
-    fn test_compile_uri_template_simple() {
-        let CompiledTemplate {
-            pattern: regex,
-            variables: vars,
-            ..
-        } = compile_uri_template("file:///{path}").unwrap();
-        assert_eq!(vars, vec!["path"]);
-        assert!(regex.is_match("file:///README.md"));
-        assert!(!regex.is_match("file:///foo/bar")); // no slashes in simple expansion
-    }
-
-    #[test]
-    fn test_compile_uri_template_multiple_vars() {
-        let CompiledTemplate {
-            pattern: regex,
-            variables: vars,
-            ..
-        } = compile_uri_template("api://v1/{resource}/{id}").unwrap();
-        assert_eq!(vars, vec!["resource", "id"]);
-        assert!(regex.is_match("api://v1/users/123"));
-        assert!(regex.is_match("api://v1/posts/abc"));
-        assert!(!regex.is_match("api://v1/users")); // missing id
-    }
-
-    #[test]
-    fn test_compile_uri_template_reserved_expansion() {
-        let CompiledTemplate {
-            pattern: regex,
-            variables: vars,
-            ..
-        } = compile_uri_template("file:///{+path}").unwrap();
-        assert_eq!(vars, vec!["path"]);
-        assert!(regex.is_match("file:///README.md"));
-        assert!(regex.is_match("file:///foo/bar/baz.txt")); // slashes allowed
-    }
-
-    #[test]
-    fn test_compile_uri_template_special_chars() {
-        let CompiledTemplate {
-            pattern: regex,
-            variables: vars,
-            ..
-        } = compile_uri_template("http://example.com/api?query={q}").unwrap();
-        assert_eq!(vars, vec!["q"]);
-        assert!(regex.is_match("http://example.com/api?query=hello"));
-    }
-
-    #[test]
-    fn test_resource_template_match_uri() {
-        let template = ResourceTemplateBuilder::new("db://users/{id}")
-            .name("User Records")
-            .handler(|uri: String, vars: HashMap<String, String>| async move {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri,
-                        mime_type: None,
-                        text: Some(format!("User {}", vars.get("id").unwrap())),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            });
-
-        // Test matching
-        let vars = template.match_uri("db://users/123").unwrap();
-        assert_eq!(vars.get("id"), Some(&"123".to_string()));
-
-        // Test non-matching
-        assert!(template.match_uri("db://posts/123").is_none());
-        assert!(template.match_uri("db://users").is_none());
-    }
-
-    #[test]
-    fn test_resource_template_match_multiple_vars() {
-        let template = ResourceTemplateBuilder::new("api://{version}/{resource}/{id}")
-            .name("API Resources")
-            .handler(|uri: String, _vars: HashMap<String, String>| async move {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri,
-                        mime_type: None,
-                        text: None,
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            });
-
-        let vars = template.match_uri("api://v2/users/abc-123").unwrap();
-        assert_eq!(vars.get("version"), Some(&"v2".to_string()));
-        assert_eq!(vars.get("resource"), Some(&"users".to_string()));
-        assert_eq!(vars.get("id"), Some(&"abc-123".to_string()));
-    }
-
-    #[tokio::test]
-    async fn test_resource_template_read() {
-        let template = ResourceTemplateBuilder::new("file:///{path}")
-            .name("Files")
-            .mime_type("text/plain")
-            .handler(|uri: String, vars: HashMap<String, String>| async move {
-                let path = vars.get("path").unwrap().clone();
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri,
-                        mime_type: Some("text/plain".to_string()),
-                        text: Some(format!("Contents of {}", path)),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            });
-
-        let vars = template.match_uri("file:///README.md").unwrap();
-        let result = template.read("file:///README.md", vars).await.unwrap();
-
-        assert_eq!(result.contents.len(), 1);
-        assert_eq!(result.contents[0].uri, "file:///README.md");
-        assert_eq!(
-            result.contents[0].text.as_deref(),
-            Some("Contents of README.md")
-        );
-    }
-
-    #[test]
-    fn test_resource_template_definition() {
-        let template = ResourceTemplateBuilder::new("db://records/{id}")
-            .name("Database Records")
-            .description("Access database records by ID")
-            .mime_type("application/json")
-            .handler(|uri: String, _vars: HashMap<String, String>| async move {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri,
-                        mime_type: None,
-                        text: None,
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            });
-
-        let def = template.definition();
-        assert_eq!(def.uri_template, "db://records/{id}");
-        assert_eq!(def.name, "Database Records");
-        assert_eq!(
-            def.description.as_deref(),
-            Some("Access database records by ID")
-        );
-        assert_eq!(def.mime_type.as_deref(), Some("application/json"));
-    }
-
-    #[test]
-    fn test_resource_template_reserved_path() {
-        let template = ResourceTemplateBuilder::new("file:///{+path}")
-            .name("Files with subpaths")
-            .handler(|uri: String, _vars: HashMap<String, String>| async move {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri,
-                        mime_type: None,
-                        text: None,
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            });
-
-        // Reserved expansion should match slashes
-        let vars = template.match_uri("file:///src/lib/utils.rs").unwrap();
-        assert_eq!(vars.get("path"), Some(&"src/lib/utils.rs".to_string()));
-    }
-
-    #[test]
-    fn test_resource_annotations() {
-        use crate::protocol::{ContentAnnotations, ContentRole};
-
-        let annotations = ContentAnnotations {
-            audience: Some(vec![ContentRole::User]),
-            priority: Some(0.8),
-            last_modified: None,
-        };
-
-        let resource = ResourceBuilder::new("file:///important.txt")
-            .name("Important File")
-            .annotations(annotations.clone())
-            .text("content");
-
-        let def = resource.definition();
-        assert!(def.annotations.is_some());
-        let ann = def.annotations.unwrap();
-        assert_eq!(ann.priority, Some(0.8));
-        assert_eq!(ann.audience.unwrap(), vec![ContentRole::User]);
-    }
-
-    #[test]
-    fn test_resource_template_annotations() {
-        use crate::protocol::{ContentAnnotations, ContentRole};
-
-        let annotations = ContentAnnotations {
-            audience: Some(vec![ContentRole::Assistant]),
-            priority: Some(0.5),
-            last_modified: None,
-        };
-
-        let template = ResourceTemplateBuilder::new("db://users/{id}")
-            .name("Users")
-            .annotations(annotations)
-            .handler(|uri: String, _vars: HashMap<String, String>| async move {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri,
-                        mime_type: None,
-                        text: Some("data".to_string()),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            });
-
-        let def = template.definition();
-        assert!(def.annotations.is_some());
-        let ann = def.annotations.unwrap();
-        assert_eq!(ann.priority, Some(0.5));
-        assert_eq!(ann.audience.unwrap(), vec![ContentRole::Assistant]);
-    }
-
-    #[test]
-    fn test_resource_no_annotations_by_default() {
-        let resource = ResourceBuilder::new("file:///test.txt")
-            .name("Test")
-            .text("content");
-
-        let def = resource.definition();
-        assert!(def.annotations.is_none());
-    }
-
-    #[test]
-    fn test_try_handler_success() {
-        let result = ResourceTemplateBuilder::new("db://users/{id}")
-            .name("Users")
-            .try_handler(|uri: String, _vars: HashMap<String, String>| async move {
-                Ok(ReadResourceResult {
-                    contents: vec![ResourceContent {
-                        uri,
-                        mime_type: None,
-                        text: Some("ok".to_string()),
-                        blob: None,
-                        meta: None,
-                    }],
-                    meta: None,
-                    ..Default::default()
-                })
-            });
-
-        assert!(result.is_ok());
-        let template = result.unwrap();
-        assert_eq!(template.uri_template, "db://users/{id}");
-    }
-
-    #[test]
-    fn test_compile_uri_template_returns_result() {
-        // Valid templates should succeed
-        assert!(compile_uri_template("file:///{path}").is_ok());
-        assert!(compile_uri_template("api://v1/{resource}/{id}").is_ok());
-        assert!(compile_uri_template("file:///{+path}").is_ok());
-        assert!(compile_uri_template("no-vars").is_ok());
-        assert!(compile_uri_template("").is_ok());
-    }
-}
+mod tests;
